Fix: Toolbar should not wrap content
Update the Toolbar widget to sit alongside content widgets instead of
wrapping them, which allows the current scroll position to be maintained
when navigating away from the screen and then returning.

Remove the Toolbar's content `div` and associated classes; place the
title, actions, and button directly beneath the root node; and remove
the toolbar node wrapper and associated classes. Also, disable setting
the position to `Position.bottom` when the `fixed` property is falsy, as
this functionality is better served by a footer component.