#1171 Convert TabController to uncontrolled widget with renderer (#1208)
Remove the `activeIndex`, `onRequestChangeTab`, and `onRequestClosedTab`
properties from `<TabController>` in favor of managing the active and
closed tabs internally. Also, add an `initialId` property so that the initial active
active tab may be set.

Update `<TabController>` to accept a required `tabs` array config
property, and replace its children with a renderer function that can
return an array of `<TabContent>` children. Also, remove `<TabButton>`
and `<Tab>` as the former has been inlined into `<TabController>` and
the latter has been replaced with the simpler `<TabContent>`.